Remove .ppd related logic and rely on CUPS to know how to print
On bullseye lpadmin complains that using drivers directly is deprecated,
hence we go with the flow and pre-emptively make sure we don't run into
this problem in the future.

This leaves printer model checks in place, but simultaneously lowers the
bar to support more printers in the future.

This change was tested with an HP LaserJet Pro M404n but *not* with the
Brother HL-L2320D